refactor(query): Split plan/groupby.rs into submodules
