+++ crates/query/src/plan/groupby/mod.rs
@@ -1,0 +1,9 @@
+//! GroupByNode for grouping query results
+
+mod node;
+mod plan_node;
+mod rendering;
+mod types;
+
+pub use node::GroupByNode;
+pub use types::{ChildSelectMeta, DocumentGroup, GroupAlias, InnerAggregateDef};
